requirements.txt
python==3.8.*
streamlit==1.0.0
Pillow==8.0.0